Move socket-reading test helpers into docker.utils.socket
